+++ components/types/index.ts
@@ -1,13 +1,11 @@
-export * from './responses'
-export * from './app.interface';
-export * from './artists.interface';
-export * from './audio-player.interface';
-export * from './favorites-dialog.interface';
-export * from './search-dialog.interface';
-export * from './sidebar.interface';
-export * from './song.interface';
-export * from './songs.interface';
-export * from './main-content.interface';
-export * from './top-nav.interface';
-export * from './main-content.interface';
-export * from './tabs.enum'
+export * from "./responses";
+export * from "./song-data.interface";
+export * from "./artists.interface";
+export * from "./audio-player.interface";
+export * from "./favorites-dialog.interface";
+export * from "./search-dialog.interface";
+export * from "./sidebar.interface";
+export * from "./song.interface";
+export * from "./songs.interface";
+export * from "./top-nav.interface";
+export * from "./tabs.enum";
