--- components/types/main-content.interface.ts
@@ -1,8 +1,0 @@
-import { TopNavProps } from './top-nav.interface';
-
-export interface MainContentProps extends TopNavProps {}
-
-export interface MainContentStates {
-    currentTab: number;
-    soundWave: boolean;
-}
