+++ components/types/sidebar.interface.ts
@@ -1,7 +1,4 @@
-import { SongFn } from "./top-nav.interface";
-
 export interface SidebarProps {
   toggle: boolean;
   closeDrawer: () => void;
-  songFn: SongFn;
 }
