+++ components/types/song-data.interface.ts
@@ -1,0 +1,5 @@
+export interface SongData {
+  path: string;
+  id: string;
+  visible: boolean;
+}
