+++ components/types/song.interface.ts
@@ -2,13 +2,9 @@ import { SongResponse } from "./responses/song-response.interface";
 import { SongFn } from "./top-nav.interface";
 
 export interface SongProps {
-    songFn: SongFn;
-    id: string;
-    song: SongResponse;
-    soundWave: boolean;
-    key: string;
-}
-
-export interface SongStates {
-    isFav: boolean;
+  songFn: SongFn;
+  id: string;
+  song: SongResponse;
+  soundWave: boolean;
+  key: string;
 }
