+++ components/types/songs.interface.ts
@@ -1,13 +1,5 @@
-import { SongFn } from './top-nav.interface';
 import { SongResponse } from "./responses/song-response.interface";
 
 export interface SongsProps {
-    songs: Array<SongResponse>
-    songFn: SongFn
-    soundWave: boolean;
-}
-
-export interface SongsStates {
-    musicPath: string;
-    visible: boolean;
+  songs: Array<SongResponse>;
 }
