--- components/types/tab-panel.interface.ts
@@ -1,8 +1,0 @@
-// react
-import { ReactNode } from "react";
-
-export interface TabPanelProps {
-    value: number;
-    index: number;
-    children: ReactNode;
-}
