+++ components/types/top-nav.interface.ts
@@ -1,20 +1,7 @@
-import { SongData } from "./app.interface";
-import { SongResponse, ArtistResponse } from "./responses";
-
-
-export interface TopNavStates {
-    toggle: boolean;
-    favOpen: boolean;
-}
-
-export interface TopNavProps {
-    songs: SongResponse[];
-    artists: ArtistResponse[];
-    songFn: SongFn;
-}
+import { SongData } from "./song-data.interface";
 
 export interface SongFn {
-    getSongPath: () => SongData;
-    setSongPath: (path: string, id: string, visible: boolean) => void;
-    openSearchDialog: () => void;
+  getSongPath: () => SongData;
+  setSongPath: (path: string, id: string, visible: boolean) => void;
+  openSearchDialog: () => void;
 }
