+++ componentsv2/styles/audio-player.style.ts
